chore/#73: type float로 변경

+++ app/src/main/java/com/wearerommies/roomie/domain/entity/SearchResultEntity.kt
@@ -1,8 +1,8 @@
 package com.wearerommies.roomie.domain.entity
 
 data class SearchResultEntity(
-    val x: Double,
-    val y: Double,
+    val x: Float,
+    val y: Float,
     val location: String,
     val address: String,
     val roadAddress: String
